requirements.txt
crewai
langchain
langchain-openai
openai
gitpython
langchain-community
